Moved get_config into config file
